Use arm instruction set
The Angstrom distro configuration enables thumb instructions support for user
applications but directFB, erlang, and possibly others don't work well with
that. It's probably related to compiler optimizations. The symptoms include runtime
errors with illegal instructions and illegal pointer operations. Since we're
building everything from the ground up there's no need to this sort of compromise.
Anyway, interworking is untouched so we may still be compatible with thumb binaries.

Signed-off-by: Nicolas Lebedenco <nicolas_lebedenco@bol.com.br>