Add kernel boot logo customization
Automates the process of translating a properly formatted png image into a
header file suitable for use as the kernel logo. In practice, it depends on netpbm
for the tools (pngtopnm, pnmquant and pnmtoplainnm) but there's no such thing as
a netpbm-native package. As of now netpbm should be installed on the host.

To create a logo file:

    1) Open the image in gimp or wherever.
    2) Resize it to 80x80 (a black background usually looks better)
    3) Save the image back to png format with then name logo.png
    4) Copy logo.png to recipes/linux/linux-mini2440-2.6.32+git/
       replacing the default one

Signed-off-by: Nicolas Lebedenco <nicolas_lebedenco@bol.com.br>